Automatic PowerUp Usage and Equip PowerUp from Locker Initial Implementation
Some notes:

alter the if(enemiesInRange.Length <= 0) case. the for loop case already handles doing nothing if there are no enemies.

delete bool powerUpEquipped and just use if(PowerUp.PowerUpType.None) instead

make the PowerUp.Use function set the player's powerUp to None, so that we don't have to keep rewriting it in the player code.

changed the HandlePowerUp function to return a bool instead of being void. The return value represents whether the powerup used on them was applicable to them. This is used to determine the automatic powerup use case.

changed pixie's returnHome function to be a void function rather than a coroutine (coroutine was resulting in multiple calls when we just want it to be called once)

created a prefab for the bug spray. I am unsure of the scaling so far in relation to the rest of the level (1, 1, 1)

created a temporary prefab for the helmet. The sprite needs to be switched out from a knob to its actual art. Same question with scaling.

changed the "Enemie" tag to be correctly spelled "Enemy"